+++ Tests/Security/UserCheckerTest.php
@@ -13,92 +13,42 @@ namespace FOS\UserBundle\Tests\Security;
 
 use FOS\UserBundle\Security\UserChecker;
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Security\Core\Exception\AccountExpiredException;
-use Symfony\Component\Security\Core\Exception\CredentialsExpiredException;
 use Symfony\Component\Security\Core\Exception\DisabledException;
-use Symfony\Component\Security\Core\Exception\LockedException;
 
 class UserCheckerTest extends TestCase
 {
-    public function testCheckPreAuthFailsLockedOut()
-    {
-        $this->expectExceptionMessage('User account is locked.');
-        $this->expectException(LockedException::class);
-
-        $userMock = $this->getUser(false, false, false, false);
-        $checker = new UserChecker();
-        $checker->checkPreAuth($userMock);
-    }
-
     public function testCheckPreAuthFailsIsEnabled()
     {
         $this->expectExceptionMessage('User account is disabled.');
         $this->expectException(DisabledException::class);
 
-        $userMock = $this->getUser(true, false, false, false);
-        $checker = new UserChecker();
-        $checker->checkPreAuth($userMock);
-    }
-
-    public function testCheckPreAuthFailsIsAccountNonExpired()
-    {
-        $this->expectExceptionMessage('User account has expired.');
-        $this->expectException(AccountExpiredException::class);
-
-        $userMock = $this->getUser(true, true, false, false);
+        $userMock = $this->getUser(false);
         $checker = new UserChecker();
         $checker->checkPreAuth($userMock);
     }
 
     public function testCheckPreAuthSuccess()
     {
-        $userMock = $this->getUser(true, true, true, false);
+        $userMock = $this->getUser(true);
         $checker = new UserChecker();
 
-        try {
-            $this->assertNull($checker->checkPreAuth($userMock));
-        } catch (\Exception $ex) {
-            $this->fail();
-        }
-    }
-
-    public function testCheckPostAuthFailsIsCredentialsNonExpired()
-    {
-        $this->expectExceptionMessage('User credentials have expired.');
-        $this->expectException(CredentialsExpiredException::class);
-
-        $userMock = $this->getUser(true, true, true, false);
-        $checker = new UserChecker();
-        $checker->checkPostAuth($userMock);
+        $this->assertNull($checker->checkPreAuth($userMock));
     }
 
     public function testCheckPostAuthSuccess()
     {
-        $userMock = $this->getUser(true, true, true, true);
+        $userMock = $this->getUser(true);
         $checker = new UserChecker();
 
-        try {
-            $this->assertNull($checker->checkPostAuth($userMock));
-        } catch (\Exception $ex) {
-            $this->fail();
-        }
+        $this->assertNull($checker->checkPostAuth($userMock));
     }
 
-    private function getUser($isAccountNonLocked, $isEnabled, $isAccountNonExpired, $isCredentialsNonExpired)
+    private function getUser($isEnabled)
     {
         $userMock = $this->getMockBuilder('FOS\UserBundle\Model\User')->getMock();
         $userMock
-            ->method('isAccountNonLocked')
-            ->willReturn($isAccountNonLocked);
-        $userMock
             ->method('isEnabled')
             ->willReturn($isEnabled);
-        $userMock
-            ->method('isAccountNonExpired')
-            ->willReturn($isAccountNonExpired);
-        $userMock
-            ->method('isCredentialsNonExpired')
-            ->willReturn($isCredentialsNonExpired);
 
         return $userMock;
     }